+++ Changelog.md
@@ -1,3 +1,3 @@
 ## v0.0.1
 
- * The command "php bin/app.php extbaser:export your_scheme target_extension_key" exports an existing database scheme to a TYPO3 Extbase Extension. This includes all tables and their properties.
+ * The command "php bin/app.php extbaser:export your_schema target_extension_key" exports an existing database schema to a TYPO3 Extbase Extension. This includes all tables and their properties.